requirements.txt
torch>=2.1.0
PyYAML>=6.0
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
seaborn>=0.13
datasets>=2.16
transformers>=4.37
tqdm>=4.66
pytest>=8.0
jupyter>=1.0
